fix segfault when cleaning up MixChunks on windows
We built the libSDL binaries several years ago, and so it was built
against an older runtime library (msvcrt.dll).
Now we build SDL itself against msvcr110.dll for example, and therefore
allocate on another heap than the free() is taking place, which happens
in libSDL_mixer itself.
See: http://stackoverflow.com/questions/23257226/sdl-mixer-mix-freechunk-crashing-on-sample-created-in-memory