clarify argments to handler subs in SDLx::App
Previously it was easy for someone to look at the docs and assume the move handler operated in fixed steps, only discovering how wrong that is if they end up with a need to dig into the SDLx::Controller documentation. With this line it becomes obvious that the move handler takes a number of arguments and that research is necessary.

And while i'm at it, also adding the arguments for the other handlers.