Allow calling SDLx::Surface draw methods with blessed vectors
Before this patch, validation of vectors used for coordinates
in SDLx::Surface draw methods checked with

    ref $vector eq 'ARRAY'

However, this meant that it was impossible to use these methods
with eg. an array-based vector class.

To support this, the check is now made with Scalar::Util::reftype,
which more accurately represents what this test is checking.